update name to GuidedSetup
